% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owd.R
\name{owd}
\alias{owd}
\title{Function opens the current working directory in your computer's file manager}
\usage{
owd(...)
}
\description{
Function opens the current working directory in your computer's file manager
}
\examples{
owd()

}
\author{
Benjamin Feakins, \email{benjamin.feakins@ndph.ox.ac.uk}
}
