\name{owd}
\alias{owd}
\title{Function to open the current working directory in your computer's file manager}
\usage{
owd()
}
\description{
Function to open the current working directory in your computer's file manager.
}
\examples{
\dontrun{owd()}
}
\author{Benjamin G. Feakins, \email{benjamin.feakins@ndph.ox.ac.uk}}
